rename head component
